pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goodshyt-core"
version = "0.1.0"
description = "Canonical ethics, metrics, protocol, and CLI foundation for the GoodShyt Group ecosystem."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.2",
  "pyyaml>=6.0.2",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.2", "httpx>=0.27.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
